refactor: 입력한 단어, 정답 단어에 대한 검증 클래스 생성
